Refactor from stiches to tailwind
