+++ apps/web/src/components/icons/FacebookIcon.tsx
@@ -1,0 +1,9 @@
+export const FacebookIcon = () => (
+  <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="0.432617" width="73.3513" height="73" rx="36.5" fill="#FAB759" />
+    <path
+      d="M53.2452 19.0435H20.9706C20.1945 19.0435 19.5674 19.6675 19.5674 20.44V52.56C19.5674 53.3324 20.1945 53.9565 20.9706 53.9565H53.2452C54.0214 53.9565 54.6485 53.3324 54.6485 52.56V20.44C54.6485 19.6675 54.0214 19.0435 53.2452 19.0435ZM49.1934 29.2337H46.3913C44.1943 29.2337 43.7689 30.2724 43.7689 31.7998V35.1646H49.0136L48.3295 40.4321H43.7689V53.9565H38.3007V40.4364H33.727V35.1646H38.3007V31.2805C38.3007 26.7723 41.0677 24.3153 45.1108 24.3153C47.049 24.3153 48.711 24.4593 49.1977 24.5248V29.2337H49.1934Z"
+      fill="#0D2546"
+    />
+  </svg>
+)
