+++ apps/web/src/components/icons/GithubIcon.tsx
@@ -1,22 +1,11 @@
-import { SVGProps } from "react"
-
-const SvgGithubIcon = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    aria-hidden="true"
-    data-prefix="fab"
-    data-icon="github"
-    className="github-icon_svg__svg-inline--fa github-icon_svg__fa-github github-icon_svg__fa-w-16"
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 496 512"
-    width="1em"
-    height="1em"
-    {...props}
-  >
+export const GitHubIcon = () => (
+  <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="0.648438" width="73.3513" height="73" rx="36.5" fill="#FAB759" />
     <path
-      fill="#fff"
-      d="M165.9 397.4c0 2-2.3 3.6-5.2 3.6-3.3.3-5.6-1.3-5.6-3.6 0-2 2.3-3.6 5.2-3.6 3-.3 5.6 1.3 5.6 3.6zm-31.1-4.5c-.7 2 1.3 4.3 4.3 4.9 2.6 1 5.6 0 6.2-2s-1.3-4.3-4.3-5.2c-2.6-.7-5.5.3-6.2 2.3zm44.2-1.7c-2.9.7-4.9 2.6-4.6 4.9.3 2 2.9 3.3 5.9 2.6 2.9-.7 4.9-2.6 4.6-4.6-.3-1.9-3-3.2-5.9-2.9zM244.8 8C106.1 8 0 113.3 0 252c0 110.9 69.8 205.8 169.5 239.2 12.8 2.3 17.3-5.6 17.3-12.1 0-6.2-.3-40.4-.3-61.4 0 0-70 15-84.7-29.8 0 0-11.4-29.1-27.8-36.6 0 0-22.9-15.7 1.6-15.4 0 0 24.9 2 38.6 25.8 21.9 38.6 58.6 27.5 72.9 20.9 2.3-16 8.8-27.1 16-33.7-55.9-6.2-112.3-14.3-112.3-110.5 0-27.5 7.6-41.3 23.6-58.9-2.6-6.5-11.1-33.3 2.6-67.9 20.9-6.5 69 27 69 27 20-5.6 41.5-8.5 62.8-8.5s42.8 2.9 62.8 8.5c0 0 48.1-33.6 69-27 13.7 34.7 5.2 61.4 2.6 67.9 16 17.7 25.8 31.5 25.8 58.9 0 96.5-58.9 104.2-114.8 110.5 9.2 7.9 17 22.9 17 46.4 0 33.7-.3 75.4-.3 83.6 0 6.5 4.6 14.4 17.3 12.1C428.2 457.8 496 362.9 496 252 496 113.3 383.5 8 244.8 8zM97.2 352.9c-1.3 1-1 3.3.7 5.2 1.6 1.6 3.9 2.3 5.2 1 1.3-1 1-3.3-.7-5.2-1.6-1.6-3.9-2.3-5.2-1zm-10.8-8.1c-.7 1.3.3 2.9 2.3 3.9 1.6 1 3.6.7 4.3-.7.7-1.3-.3-2.9-2.3-3.9-2-.6-3.6-.3-4.3.7zm32.4 35.6c-1.6 1.3-1 4.3 1.3 6.2 2.3 2.3 5.2 2.6 6.5 1 1.3-1.3.7-4.3-1.3-6.2-2.2-2.3-5.2-2.6-6.5-1zm-11.4-14.7c-1.6 1-1.6 3.6 0 5.9 1.6 2.3 4.3 3.3 5.6 2.3 1.6-1.3 1.6-3.9 0-6.2-1.4-2.3-4-3.3-5.6-2z"
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M37.3242 19.0435C27.6331 19.0435 19.7837 27.0528 19.7837 36.9415C19.7837 44.8614 24.8047 51.5508 31.777 53.9223C32.6541 54.0789 32.9829 53.542 32.9829 53.0722C32.9829 52.6471 32.961 51.2376 32.961 49.7386C28.554 50.5664 27.4138 48.6424 27.063 47.6356C26.8657 47.121 26.0106 45.5326 25.2651 45.1075C24.6512 44.7719 23.7742 43.9441 25.2432 43.9218C26.6245 43.8994 27.6112 45.2194 27.94 45.7563C29.5187 48.4634 32.0401 47.7027 33.0487 47.2329C33.2022 46.0695 33.6626 45.2865 34.1669 44.839C30.2642 44.3916 26.186 42.8479 26.186 36.0019C26.186 34.0555 26.8657 32.4446 27.9839 31.1918C27.8085 30.7443 27.1946 28.9098 28.1593 26.4488C28.1593 26.4488 29.6283 25.979 32.9829 28.2833C34.3862 27.8806 35.8771 27.6793 37.3681 27.6793C38.859 27.6793 40.35 27.8806 41.7532 28.2833C45.1078 25.9566 46.5768 26.4488 46.5768 26.4488C47.5416 28.9098 46.9277 30.7443 46.7523 31.1918C47.8705 32.4446 48.5502 34.0331 48.5502 36.0019C48.5502 42.8703 44.4501 44.3916 40.5473 44.839C41.1831 45.3984 41.7313 46.4722 41.7313 48.1502C41.7313 50.5441 41.7093 52.4681 41.7093 53.0722C41.7093 53.542 42.0382 54.1013 42.9153 53.9223C46.3973 52.7227 49.4231 50.4392 51.5666 47.393C53.7102 44.3469 54.8637 40.6916 54.8647 36.9415C54.8647 27.0528 47.0154 19.0435 37.3242 19.0435Z"
+      fill="#0D2546"
     />
   </svg>
 )
-
-export default SvgGithubIcon
