+++ apps/web/src/components/icons/InstagramIcon.tsx
@@ -1,0 +1,9 @@
+export const InstagramIcon = () => (
+  <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="0.216309" width="73.3513" height="73" rx="36.5" fill="#FAB759" />
+    <path
+      d="M36.8874 30.6787C33.6666 30.6787 31.0382 33.2945 31.0382 36.5C31.0382 39.7054 33.6666 42.3213 36.8874 42.3213C40.1082 42.3213 42.7366 39.7054 42.7366 36.5C42.7366 33.2945 40.1082 30.6787 36.8874 30.6787ZM54.4307 36.5C54.4307 34.0894 54.4526 31.7006 54.3166 29.2943C54.1806 26.4994 53.5399 24.0189 51.4863 21.9751C49.4283 19.9269 46.9403 19.2937 44.132 19.1583C41.7098 19.0229 39.3096 19.0448 36.8918 19.0448C34.4696 19.0448 32.0693 19.0229 29.6515 19.1583C26.8432 19.2937 24.3508 19.9313 22.2972 21.9751C20.2392 24.0232 19.603 26.4994 19.467 29.2943C19.3309 31.7049 19.3529 34.0937 19.3529 36.5C19.3529 38.9062 19.3309 41.2994 19.467 43.7057C19.603 46.5006 20.2436 48.9811 22.2972 51.0249C24.3552 53.0731 26.8432 53.7063 29.6515 53.8417C32.0737 53.977 34.474 53.9552 36.8918 53.9552C39.314 53.9552 41.7142 53.977 44.132 53.8417C46.9403 53.7063 49.4327 53.0687 51.4863 51.0249C53.5443 48.9767 54.1806 46.5006 54.3166 43.7057C54.457 41.2994 54.4307 38.9106 54.4307 36.5ZM36.8874 45.4569C31.907 45.4569 27.8876 41.4566 27.8876 36.5C27.8876 31.5433 31.907 27.5431 36.8874 27.5431C41.8678 27.5431 45.8872 31.5433 45.8872 36.5C45.8872 41.4566 41.8678 45.4569 36.8874 45.4569ZM46.2558 29.2681C45.093 29.2681 44.1539 28.3335 44.1539 27.1763C44.1539 26.019 45.093 25.0844 46.2558 25.0844C47.4186 25.0844 48.3577 26.019 48.3577 27.1763C48.358 27.4511 48.3039 27.7232 48.1984 27.9772C48.0929 28.2311 47.9381 28.4619 47.7428 28.6562C47.5476 28.8505 47.3157 29.0046 47.0606 29.1096C46.8054 29.2146 46.5319 29.2684 46.2558 29.2681Z"
+      fill="#0D2546"
+    />
+  </svg>
+)
