+++ apps/web/src/components/icons/SlackIcon.tsx
@@ -1,0 +1,29 @@
+export const SlackIcon = () => (
+  <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect width="73.3513" height="73" rx="36.5" fill="#FAB759" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M33.5915 19.0435C32.6611 19.0437 31.7689 19.4117 31.1111 20.0666C30.4534 20.7215 30.084 21.6096 30.0842 22.5356C30.0844 23.4615 30.4542 24.3495 31.1122 25.004C31.7703 25.6586 32.6627 26.0263 33.5931 26.0261H37.1012V22.5348C37.1004 21.6091 36.7305 20.7215 36.0728 20.067C35.415 19.4124 34.5232 19.0443 33.5931 19.0435H33.5915ZM33.5915 28.3541H24.2376C23.3072 28.3541 22.4149 28.722 21.757 29.3767C21.0991 30.0315 20.7295 30.9195 20.7295 31.8454C20.7295 32.7714 21.0991 33.6594 21.757 34.3142C22.4149 34.9689 23.3072 35.3367 24.2376 35.3367H33.5915C34.5219 35.3367 35.4142 34.9689 36.0721 34.3142C36.73 33.6594 37.0996 32.7714 37.0996 31.8454C37.0996 30.9195 36.73 30.0315 36.0721 29.3767C35.4142 28.722 34.5219 28.3541 33.5915 28.3541Z"
+      fill="#0D2546"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M55.8107 31.8438C55.8107 30.9179 55.4411 30.0299 54.7832 29.3751C54.1253 28.7204 53.233 28.3525 52.3026 28.3525C51.3721 28.3525 50.4798 28.7204 49.8219 29.3751C49.1641 30.0299 48.7944 30.9179 48.7944 31.8438V35.3351H52.3026C53.233 35.3351 54.1253 34.9673 54.7832 34.3126C55.4411 33.6578 55.8107 32.7698 55.8107 31.8438ZM46.4552 31.8438V22.5348C46.4552 21.6088 46.0856 20.7208 45.4277 20.066C44.7698 19.4113 43.8775 19.0435 42.9471 19.0435C42.0167 19.0435 41.1244 19.4113 40.4665 20.066C39.8086 20.7208 39.439 21.6088 39.439 22.5348V31.8438C39.439 32.7698 39.8086 33.6578 40.4665 34.3126C41.1244 34.9673 42.0167 35.3351 42.9471 35.3351C43.8775 35.3351 44.7698 34.9673 45.4277 34.3126C46.0856 33.6578 46.4552 32.7698 46.4552 31.8438Z"
+      fill="#0D2546"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M42.9471 53.9565C43.8775 53.9565 44.7698 53.5887 45.4277 52.9339C46.0856 52.2792 46.4552 51.3911 46.4552 50.4652C46.4552 49.5392 46.0856 48.6512 45.4277 47.9965C44.7698 47.3417 43.8775 46.9739 42.9471 46.9739H39.439V50.4652C39.4374 52.3902 41.008 53.9533 42.9471 53.9565ZM42.9471 44.6442H52.3026C53.233 44.6442 54.1253 44.2764 54.7832 43.6216C55.4411 42.9669 55.8107 42.0789 55.8107 41.1529C55.8107 40.227 55.4411 39.3389 54.7832 38.6842C54.1253 38.0295 53.233 37.6616 52.3026 37.6616H42.9487C42.0183 37.6614 41.1259 38.029 40.4678 38.6836C39.8098 39.3382 39.44 40.2262 39.4398 41.1521C39.4396 42.0781 39.809 42.9662 40.4667 43.6211C41.1245 44.276 42.0167 44.644 42.9471 44.6442Z"
+      fill="#0D2546"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M20.7295 41.1545C20.7295 42.0805 21.0991 42.9685 21.757 43.6232C22.4149 44.278 23.3072 44.6458 24.2376 44.6458C25.168 44.6458 26.0603 44.278 26.7182 43.6232C27.3761 42.9685 27.7457 42.0805 27.7457 41.1545V37.6632H24.2376C23.3072 37.6632 22.4149 38.031 21.757 38.6858C21.0991 39.3405 20.7295 40.2286 20.7295 41.1545ZM30.085 41.1545V50.4636C30.0848 51.3896 30.4542 52.2777 31.1119 52.9326C31.7697 53.5875 32.6619 53.9555 33.5923 53.9557C34.5227 53.9559 35.4151 53.5883 36.0731 52.9337C36.7312 52.2791 37.101 51.3911 37.1012 50.4652V41.1561C37.1014 40.2301 36.732 39.342 36.0743 38.6871C35.4165 38.0322 34.5243 37.6642 33.5939 37.664C32.6635 37.6638 31.7711 38.0314 31.113 38.686C30.455 39.3406 30.0852 40.2286 30.085 41.1545Z"
+      fill="#0D2546"
+    />
+  </svg>
+)
