+++ apps/web/src/components/organisms/Footer/components/animation.css
