+++ apps/web/src/components/organisms/Footer/sections/ContactSection.tsx
@@ -1,0 +1,11 @@
+export const ContactSection = () => (
+  <div className="flex flex-col text-center">
+    <div className="text-slate-12 font-medium">Feil på nettsiden?</div>
+    <div className="text-slate-12 font-medium">
+      Ta kontakt med{" "}
+      <a className="text-red-11 font-medium" href="mailto:dotkom@online.ntnu.no">
+        Dotkom
+      </a>
+    </div>
+  </div>
+)
