--- src/components/icons/GitHubIcon.tsx
@@ -1,16 +1,0 @@
-import { VFC } from "react";
-import { Icon } from "./Icon";
-
-export const GitHubIcon: VFC = () => (
-  <Icon>
-    <svg viewBox="0 0 74 73" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect x="0.648438" width="73.3513" height="73" rx="36.5" fill="#FAB759" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M37.3242 19.0435C27.6331 19.0435 19.7837 27.0528 19.7837 36.9415C19.7837 44.8614 24.8047 51.5508 31.777 53.9223C32.6541 54.0789 32.9829 53.542 32.9829 53.0722C32.9829 52.6471 32.961 51.2376 32.961 49.7386C28.554 50.5664 27.4138 48.6424 27.063 47.6356C26.8657 47.121 26.0106 45.5326 25.2651 45.1075C24.6512 44.7719 23.7742 43.9441 25.2432 43.9218C26.6245 43.8994 27.6112 45.2194 27.94 45.7563C29.5187 48.4634 32.0401 47.7027 33.0487 47.2329C33.2022 46.0695 33.6626 45.2865 34.1669 44.839C30.2642 44.3916 26.186 42.8479 26.186 36.0019C26.186 34.0555 26.8657 32.4446 27.9839 31.1918C27.8085 30.7443 27.1946 28.9098 28.1593 26.4488C28.1593 26.4488 29.6283 25.979 32.9829 28.2833C34.3862 27.8806 35.8771 27.6793 37.3681 27.6793C38.859 27.6793 40.35 27.8806 41.7532 28.2833C45.1078 25.9566 46.5768 26.4488 46.5768 26.4488C47.5416 28.9098 46.9277 30.7443 46.7523 31.1918C47.8705 32.4446 48.5502 34.0331 48.5502 36.0019C48.5502 42.8703 44.4501 44.3916 40.5473 44.839C41.1831 45.3984 41.7313 46.4722 41.7313 48.1502C41.7313 50.5441 41.7093 52.4681 41.7093 53.0722C41.7093 53.542 42.0382 54.1013 42.9153 53.9223C46.3973 52.7227 49.4231 50.4392 51.5666 47.393C53.7102 44.3469 54.8637 40.6916 54.8647 36.9415C54.8647 27.0528 47.0154 19.0435 37.3242 19.0435Z"
-        fill="#0D2546"
-      />
-    </svg>
-  </Icon>
-);
