--- src/components/icons/Icon.tsx
@@ -1,9 +1,0 @@
-import { styled } from "@theme";
-
-export const Icon = styled("figure", {
-  width: "55px",
-  height: "55px",
-  display: "inline-block",
-  cursor: "pointer",
-  margin: "0px 13px 0px 13px",
-});
