--- src/components/organisms/Footer/components/FooterSection.tsx
@@ -1,17 +1,0 @@
-import { styled } from "@theme"
-
-export const FooterSection = styled("div", {
-  textAlign: "center",
-  margin: "auto",
-
-  variants: {
-    marginSize: {
-      small: {
-        marginBottom: "0.5em",
-      },
-      medium: {
-        marginBottom: "1.8em",
-      },
-    },
-  },
-})
