--- src/components/organisms/Footer/components/Waves.tsx
@@ -1,31 +1,0 @@
-import { styled } from "@theme"
-import { VFC } from "react"
-import "./animation.css"
-
-const Wave = styled("svg", {
-  width: "100%",
-  //marginBottom: "-10px",
-  margin: 0,
-  padding: "0",
-  height: "100px",
-})
-
-const ParallaxContainer = styled("g", {
-  //margin: "0",
-  //padding: "0",
-  //transform: "translateY(180px)",
-})
-
-export const Waves: VFC = () => (
-  <Wave viewBox="0 25 100 35" preserveAspectRatio="none" shapeRendering="auto">
-    <defs>
-      <path id="gentle-wave" d="M-0 44c30 0 58-18 88-18s 58 18 88 18 58-18 88-18 58 18 88 18 v44h-352z"></path>
-    </defs>
-    <ParallaxContainer className="parallax">
-      <use href="#gentle-wave" x="0" y="0" fill="rgba(13, 37, 70, 0.4)"></use>
-      <use href="#gentle-wave" x="-10" y="5" fill="rgba(13, 37, 70, 0.4)"></use>
-      <use href="#gentle-wave" x="-20" y="10" fill="rgba(13, 37, 70, 0.4)"></use>
-      <use href="#gentle-wave" x="-30" y="15" fill="rgba(13, 37, 70, 1)"></use>
-    </ParallaxContainer>
-  </Wave>
-)
